Separate locations steps. No shared shouty.
