+++ behat.yml
@@ -3,3 +3,4 @@ default:
     default:
       contexts:
         - FeatureContext
+        - LocationContext
